% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gbeta1.R
\name{fitMcGBB}
\alias{fitMcGBB}
\title{Fitting the McDonald Generalized beta  binomial distribution when binomial
random variable, frequency and shape parameters are given}
\usage{
fitMcGBB(x,obs.freq,a,b,c,print)
}
\arguments{
\item{x}{vector of binomial random variables}

\item{obs.freq}{vector of frequencies}

\item{a}{single value for shape parameter alpha representing a}

\item{b}{single value for shape parameter beta representing b}

\item{c}{single value for shape parameter gamma representing c}

\item{print}{logical value for print or not}
}
\value{
The output of \code{fitGHGBB} gives a list format consisting

\code{bin.ran.var} binomial random variables

\code{obs.freq} corresponding observed frequencies

\code{exp.freq} corresponding expected frequencies

\code{statistic} chi-squared test statistics

\code{df} degree of freedom

\code{p.value} probability value by chi-squared test statistic

\code{over.dis.para} over dispersion value.
}
\description{
The function will fit the McDonald Generalized Beta  Binomial Distribution
when random variables, corresponding frequencies and shape parameters are given. It will provide
the expected frequencies, chi-squared test statistics value, p value, degree of freedom
and over dispersion value so that it can be seen if this distribution fits the data.
}
\details{
\deqn{0 < a,b,c}
\deqn{x = 0,1,2,...}
\deqn{obs.freq \ge 0}

\strong{NOTE} : If input parameters are not in given domain conditions necessary
error messages will be provided to go further.
}
\examples{
No.D.D=0:7       #assigning the random variables
Obs.fre.1=c(47,54,43,40,40,41,39,95)          #assigning the corresponding frequencies
#estimating the parameters using maximum log likelihood value and assigning it
\dontrun{
parameters=suppressWarnings(bbmle::mle2(EstMLEMcGBB,start = list(a=0.1,b=0.1,c=0.2),
data = list(x=No.D.D,freq=Obs.fre.1)))
aMcGBB=bbmle::coef(parameters)[1]         #assigning the estimated a
bMcGBB=bbmle::coef(parameters)[2]         #assigning the estimated b
cMcGBB=bbmle::coef(parameters)[3]         #assigning the estimated c

#fitting when the random variable,frequencies,shape parameter values are given.
fitMcGBB(No.D.D,Obs.fre.1,aMcGBB,bMcGBB,cMcGBB)
#extracting the expected frequencies
fitMcGBB(No.D.D,Obs.fre.1,aMcGBB,bMcGBB,cMcGBB,FALSE)$exp.freq
}
}
\references{
Manoj, C., Wijekoon, P. & Yapa, R.D., 2013. The McDonald Generalized Beta-Binomial Distribution: A New
Binomial Mixture Distribution and Simulation Based Comparison with Its Nested Distributions in Handling
Overdispersion. International Journal of Statistics and Probability, 2(2), pp.24-41.

Available at: \url{http://www.ccsenet.org/journal/index.php/ijsp/article/view/23491} .

Janiffer, N.M., Islam, A. & Luke, O., 2014. Estimating Equations for Estimation of Mcdonald Generalized
Beta - Binomial Parameters. , (October), pp.702-709.

Roozegar, R., Tahmasebi, S. & Jafari, A.A., 2015. The McDonald Gompertz Distribution: Properties and Applications.
Communications in Statistics - Simulation and Computation, (May), pp.0-0.

Available at: \url{http://www.tandfonline.com/doi/full/10.1080/03610918.2015.1088024} .
}
\seealso{
\code{\link[bbmle]{mle2}}
}
